Place SI object immediatelly after the MOT directory object
TS 102 371 recommends that SI object is transmitted immediatelly after every transmission of the MOT directory object. In such a case when any MOT object with a logo is received, the receiver already knows, what to do with it.

SI object is now placed at the beginning of the objects list instead of at the end.