Include all DAB bearers for a service in SI
Reverts Commit 067c16c.

TS 102 371 sais "the bearer elements with a domain that matches the delivery system shall be encoded"

Encoding all DAB bearers has advantages for the user. Logos can be seen even if the other bearer does not transmit SPI. And without the need to wait for their reception when the receiver changes the MUX. Few bytes of additional payload is very low price for the added value.